+++ src/main/java/kr/co/hrmanager/domain/departments/DepartmentsRepository.java
@@ -2,4 +2,4 @@ package kr.co.hrmanager.domain.departments;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DepartmentsRepository  extends JpaRepository<Departments, Integer>  { }
+public interface DepartmentsRepository  extends JpaRepository<Departments, Long>  { }
